Update 20190712065651_add_column_to_estimates.rb
Changed design type from integer to float

+++ db/migrate/20190712065651_add_column_to_estimates.rb
@@ -1,6 +1,6 @@
 class AddColumnToEstimates < ActiveRecord::Migration[5.2]
   def change
-    add_column :estimates, :design, :integer
+    add_column :estimates, :design, :float
     add_column :estimates, :qaTesting, :float
     add_column :estimates, :deployment, :float
     add_column :estimates, :postDeploymentDev, :float
